requirements.txt
tiktoken
pyautogen
arxiv
chromadb
pypdf
langchain
openai-functools